fixing tests and splitting out hours of work

+++ src/main/scala/models/work_time/HoursOfWork.scala
@@ -6,7 +6,7 @@ import io.circe.generic.semiauto.deriveDecoder
 import io.circe.generic.semiauto.deriveEncoder
 
 case class HoursOfWork(
-    hoursOfWork: Double                    
+    hoursOfWork: BigDecimal                    
 )
 
 object HoursOfWork {
